feat: Add message field to applications table and default value to `status` column
- Add message field to application table for storing the content of the application
- Add default value (`pending`) to status column of the applications table

This commit adds a new `message` field to the `applications` table in the database. The field is of type `TEXT` and is marked as `NOT NULL`. Additionally, a default value of 'pending' is set for the `status` field. This change allows for storing a message along with the application and provides a default status for new applications.