feat(user): Add role field to users table and User model
The commit adds a `role` field to the `users` table in the database and the `User` model in the `user.rs` file. The field is of type `TEXT` and is marked as `NOT NULL`. This change allows for storing the role of a user and enhances the user management capabilities of the application.